refactor some methods
